+++ core/util/include/util/matrix.h
@@ -13,16 +13,16 @@
 
 typedef struct MyMatrix {
     unsigned int r, c;
-    double** matrix;
+    float** matrix;
 
     bool transposed; // todo
 } mymatrix_t;
 
 mymatrix_t matrix_allocator(int r, int c);
-mymatrix_t matrix_constructor(int r, int c, double** matrix);
+mymatrix_t matrix_constructor(int r, int c, float** matrix);
 
 /**
- * Frees the allocated 2d double array of the matrix
+ * Frees the allocated 2d float array of the matrix
  */
 void matrix_free(mymatrix_t m);
 mymatrix_t matrix_copy(mymatrix_t src);
@@ -31,7 +31,7 @@ void matrix_memcpy(mymatrix_t dst, mymatrix_t src);
 void matrix_multiply(mymatrix_t m1, mymatrix_t m2, 
                      mymatrix_t result);
 
-void matrix_multiply_scalar(mymatrix_t m1, double scalar,
+void matrix_multiply_scalar(mymatrix_t m1, float scalar,
                             mymatrix_t result);
 
 void matrix_elementwise_multiply(mymatrix_t m1, mymatrix_t m2,
@@ -55,11 +55,11 @@ void matrix_transpose(mymatrix_t m,
 
 bool matrix_equal(mymatrix_t m1, mymatrix_t m2);
 
-void matrix_for_each_operator(mymatrix_t m, double (*op)(double), 
+void matrix_for_each_operator(mymatrix_t m, float (*op)(float), 
                               mymatrix_t result);
 
 void matrix_print(mymatrix_t m);
 
-void matrix_set_values_to_fit(mymatrix_t m, double* elements, unsigned int num_elements);
+void matrix_set_values_to_fit(mymatrix_t m, float* elements, unsigned int num_elements);
 
 #endif // MATRIX_H
